fix exposed settings functions

+++ controllers/admin/settings.php
@@ -2,6 +2,8 @@
 
 namespace Admin;
 
+use Base;
+
 class Settings extends AdminController {
 
 	public function index($f3) {
@@ -22,12 +24,14 @@ class Settings extends AdminController {
 	}
 
 	public function clearcache($f3) {
+		if(!defined('DEBUG')) return Base::instance()->reroute('/admin/settings'); 
 		$cache = isset($this->request->data['cache']) ? getcwd() . '/' . $this->request->data['cache'] : getcwd() . '/tmp/cache';
 		$cache = str_replace(".","",$cache);
 		$this->delTree($cache);
 	}
 
-	public function delTree($dir) { 
+	public function delTree($dir) {
+		if(!defined('DEBUG')) return Base::instance()->reroute('/admin/settings'); 
 		$files = array_diff(scandir($dir), array('.','..')); 
 		foreach ($files as $file) {
 			(is_dir("$dir/$file") && !is_link($dir)) ? $this->delTree("$dir/$file") : unlink("$dir/$file");
